+++ backends/carp_webservices/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.20.1
+* support for more file endpoints
+   * `getAllFiles()` - gettting all files in a study
+   * `queryFiles()` - query for specific files in a study
+* support for uploading and downloading informed consent to be shown to the user
+
 ## 0.20.0
 * **BREAKING**: Now using the [`carp_core`](https://pub.dev/packages/carp_core) domain models in order to align with the overall [domain-driven design of CARP](https://carp.cachet.dk/core/)
 * upgrade to `carp_mobile_sensing` v. 0.20.x